// src/redux/orderMKP/types.tsx
export enum ReturnOrderActionTypes {
  // Search
  RETURN_ORDER_SEARCH_REQ = '@RETURN_ORDER/SEARCH_REQ',
  RETURN_ORDER_SEARCH_SUCCESS = '@RETURN_ORDER/SEARCH_SUCCESS',
  RETURN_ORDER_SEARCH_FAIL = '@RETURN_ORDER/SEARCH_FAIL',

  // Create
  RETURN_ORDER_CREATE_REQ = '@RETURN_ORDER/CREATE_REQ',
  RETURN_ORDER_CREATE_SUCCESS = '@RETURN_ORDER/CREATE_SUCCESS',
  RETURN_ORDER_CREATE_FAIL = '@RETURN_ORDER/CREATE_FAIL',

  // Update SR
  RETURN_ORDER_UPDATE_SR_REQ = '@RETURN_ORDER/UPDATE_SR_REQ',
  RETURN_ORDER_UPDATE_SR_SUCCESS = '@RETURN_ORDER/UPDATE_SR_SUCCESS',
  RETURN_ORDER_UPDATE_SR_FAIL = '@RETURN_ORDER/UPDATE_SR_FAIL',

  // Update Status/Confirm
  RETURN_ORDER_UPDATE_STATUS_REQ = '@RETURN_ORDER/UPDATE_STATUS_REQ',
  RETURN_ORDER_UPDATE_STATUS_SUCCESS = '@RETURN_ORDER/UPDATE_STATUS_SUCCESS',
  RETURN_ORDER_UPDATE_STATUS_FAIL = '@RETURN_ORDER/UPDATE_STATUS_FAIL',

  // Cancel
  RETURN_ORDER_CANCEL_REQ = '@RETURN_ORDER/CANCEL_REQ',
  RETURN_ORDER_CANCEL_SUCCESS = '@RETURN_ORDER/CANCEL_SUCCESS',
  RETURN_ORDER_CANCEL_FAIL = '@RETURN_ORDER/CANCEL_FAIL',

  // Mark Edited
  RETURN_ORDER_MARK_EDITED_REQ = '@RETURN_ORDER/MARK_EDITED_REQ',
  RETURN_ORDER_MARK_EDITED_SUCCESS = '@RETURN_ORDER/MARK_EDITED_SUCCESS',
  RETURN_ORDER_MARK_EDITED_FAIL = '@RETURN_ORDER/MARK_EDITED_FAIL',

  // Reset
  RETURN_ORDER_RESET = '@RETURN_ORDER/RESET'
}